Refactor fallback loader-related logic for reusability.
