+++ module/VuFind/src/VuFind/RecordDriver/Summon.php
@@ -27,6 +27,8 @@
  */
 namespace VuFind\RecordDriver;
 
+use VuFind\RecordDriver\Feature\PreviousUniqueIdInterface;
+
 /**
  * Model for Summon records.
  *
@@ -36,8 +38,10 @@ namespace VuFind\RecordDriver;
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development:plugins:record_drivers Wiki
  */
-class Summon extends DefaultRecord
+class Summon extends DefaultRecord implements PreviousUniqueIdInterface
 {
+    use \VuFind\RecordDriver\Feature\PreviousUniqueIdTrait;
+
     /**
      * Fields that may contain subject headings, and their descriptions
      *
@@ -56,35 +60,6 @@ class Summon extends DefaultRecord
      * @var \VuFind\Date\Converter
      */
     protected $dateConverter = null;
-
-    /**
-     * Previous unique ID (if applicable).
-     *
-     * @var string
-     */
-    protected $previousUniqueId = null;
-
-    /**
-     * Get previous unique ID (or null if not applicable).
-     *
-     * @return string
-     */
-    public function getPreviousUniqueId()
-    {
-        return $this->previousUniqueId;
-    }
-
-    /**
-     * Set previous unique ID
-     *
-     * @param string $id ID to set
-     *
-     * @return void
-     */
-    public function setPreviousUniqueId($id)
-    {
-        $this->previousUniqueId = $id;
-    }
 
     /**
      * Get all subject headings associated with this record.  Each heading is
